:sparkles: add extracted components for Dialog* items
